requirements.txt
langchain
langchain-text-splitters
langchain-community
langgraph

chromadb
langchain-chroma

langchain-groq
langchain-google-genai

python-dotenv

rank_bm25

ragas
datasets

deepeval